requirements.txt
APScheduler==3.5.3
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
Flask==1.0.2
google-api-python-client==1.7.4
google-auth==1.6.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.11.3
idna==2.7
oauthlib==2.1.0
peewee==3.7.1
praw==6.0.0
prawcore==1.0.0
psycopg2==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.0.0
rsa==4.0
six==1.11.0
update-checker==0.16
uritemplate==3.0.0
urllib3==1.24.1
newrelic
